+++ CHANGELOG.md
@@ -1,9 +1,7 @@
 # Changelog
 
 ## v2.0.0-beta.3
-- HTTP API routes:
-  - `GET /api/database`
-  - `GET /database/:collectionName`
+- Database library REST API
 - New build system.
 - New application startup system.